requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==3.4.7
Django==3.2.5
django-fernet-fields==0.6
djangorestframework==3.12.4
pycparser==2.20
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
